SFR-1432 EB to ECS
This migrates the CCE backend from an ElasticBeanstalk deployment to a containerized approach. This is the first cut, removing EB-specific code and making the application easier to run in a generic way. This also allows for the reading of a config YAML file and for 2 modes of operation: `api` and `pipeline`.