HOTFIX Update ES reindexing point
Previously the pipeline process used the `loadFromTime` variable to determine starting point for ElasticSearch indexing. This had unintended side-effects of forcing a reindex of ALL records if that value was not set. This change sets it so that the reindexing only updates records with a `date_modified` later than the start time of the pipeline execution